requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pyaes==1.6.1
pyasn1==0.4.8
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.3.1
python-binance==1.0.15
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.27.1
rsa==4.8
six==1.16.0
TgCrypto==1.2.5
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
